Allow the values to be in hex format (0x/0X).
E.g. spirit:AddressOffset and spirit:reset can now be specified in hex.

If the Python built-in function 'int' has a value 0 for the second
argument, it will discover the base of the first argument.

Note this is not 100% the same as specified in the IP-XACT schema
(simpleTypes.xsd).  E.g. the magnitude (K, M, ...) is not supported.

Signed-off-by: Jan Vermaete <jan.vermaete@gmail.com>