Some work on the enumerations.
Added the enumeration description to the VHDL code as comment.

Moved the info of the enumerations in the generated documentation
from the table with the fields in a register to it's own table.
This allowed to add the description of the enumeration values.
And make it more readable when having long names for enums.

The generated RTL code is still functional the same.
Signed-off-by: Jan Vermaete <jan.vermaete@gmail.com>