rst and md: Use the name of the addressBlock as title.
A design can have multiple IP's with registers.
And although IP-XACT can handle this, the tool right now only starts at the
addressBlock level.
But when collecting the generated documentation (restucturedText or MarkDown)
into one document is useful to have the name of the IP as the title of the
generated file.

This fixes issue #43: ipxact2hdlCommon.py returnRstTitle returns hardcoded string iso addressBlock.name

Signed-off-by: Jan Vermaete <jan.vermaete@gmail.com>