vhdl: generate std_logic or std_ulogic logic
Default the tool will generate unresolved logic (std_ulogic & std_ulogic_vector).
But with setting 'std' in the config file under 'vhdl' at 'resolved', the generated
VHDL code will be with std_logic.

note: the naming in the python code of this tool will still use 'sulv' in the python
      function names for historical reasons..

Signed-off-by: Jan Vermaete <jan.vermaete@gmail.com>