Fix regression caused by haedeea741b1b32d4cb79416da276f24974600949:
Addressed an issue where fields of a register were not ordered by bit offset in the IPXACT XML file.
This ordering problem caused the documentation of the fields to be twised.
And a small clean-up of whitespaces.

Swapped the fields in the first register of the test2.xml file to test this case.

Signed-off-by: Jan Vermaete <jan.vermaete@gmail.com>