Sphinx: Make the RST references to register and enum unique.
The RestructuredText references to register and enumarations of fields
should be unique.

In IPXact the name of a register can be the same as the name of a field.
But in sphinx they must have an uniq name.

So go from:

    AAAA
    ----

    a register

    AAAA
    ~~~~

    Description of e.q. the enumeration of a field.

To:

    .. _reg_AAAA:

    AAAA
    ----

    .. _field_AAAA:

    AAAA
    ~~~~

Without this Sphinx is give an error/warning because the link to a register
is used in the overview table of the registers.

Signed-off-by: Jan Vermaete <jan.vermaete@gmail.com>